refactor: user 관련 추가 리팩토링
