+++ src/utils/constants/user.ts
@@ -1,0 +1,24 @@
+export const USER_URL = {
+  GET_ALL_USER: '/users/get-users',
+  GET_SEARCHED_USER: '/search/users/',
+  GET_CLICKED_USER: '/users/',
+};
+
+export const FILE = {
+  JPEG: 'image/jpeg',
+  PNG: 'image/png',
+  GIF: 'image/gif',
+  SVG: 'image/svg+xml',
+};
+
+export const FORM_DATA = {
+  POST_ID: 'postId',
+  IMAGE: 'image',
+  TITLE: 'title',
+  CHANNEL_ID: 'channelId',
+};
+
+export const ERROR_MESSAGE = {
+  UPLOAD: '사진을 업로드 해주세요',
+  UPLOAD_IMAGE: '이미지 파일만 업로드 가능합니다',
+};
